// src/api/models/CreateFlightRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateFlightRequest = {
    flightNumber?: string;
    airlineCode?: string;
    airlineName?: string;
    departureAirportCode?: string;
    departureTime?: string;
    departureTerminal?: string;
    arrivalAirportCode?: string;
    arrivalTime?: string;
    arrivalTerminal?: string;
    aircraftType?: string;
    totalSeats?: number;
    basePrice?: number;
};