// src/api/models/AirportResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
 
export type AirportResponse = {
    code?: string;
    name?: string;
    displayName?: string;
    city?: string;
    country?: string;
    location?: string;
    timeZone?: string;
};